requirements.txt
# Google Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# HTTP & Networking
requests==2.31.0

# Scheduling & Background Jobs
APScheduler==3.10.4

# Timezone
pytz==2023.3

# System Monitoring
psutil==7.1.0

# Database (Supabase)
supabase==2.3.4
postgrest==0.16.0

# Security & Encryption
bcrypt==4.1.2
cryptography==42.0.2
